bootcamp-solutions/4-apache-flink-training/homework/requirements.txt
kafka-python==2.0.2
python-dateutil==2.8.2
psycopg2-binary==2.9.1
requests
faker